data_wave/backend/scripts_automation/requirements.txt
# FastAPI and web server
fastapi==0.104.1
pydantic==1.10.13
uvicorn[standard]==0.24.0

# Database - CRITICAL: Fixed versions for enterprise production
sqlmodel==0.0.14
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
pymysql>=1.0.3
pymongo>=4.3.3
mysql-connector-python>=8.0.0
snowflake-connector-python>=3.0.0
boto3>=1.26.0
redis>=4.5.0
azure-identity>=1.12.0

# Utilities
python-dotenv>=1.0.0
requests>=2.28.2
croniter>=1.4.1
asyncio>=3.4.3

# Data processing - CRITICAL: Fixed versions for compatibility
numpy==1.24.3
pandas==1.5.3

# Authentication and security
python-jose[cryptography]
passlib[bcrypt]
apscheduler
cryptography>=40.0.0
python-multipart
pyotp==2.8.0
alembic==1.11.3
aiosmtplib
msal

# Testing
pytest>=7.3.1
httpx>=0.24.0

# Machine Learning - CRITICAL: Fixed versions for enterprise production
scikit-learn==1.3.0
joblib==1.3.2
matplotlib==3.7.2
xgboost==1.7.6
Authlib

# ========================================
# ENTERPRISE DATA GOVERNANCE DEPENDENCIES
# ========================================

# Advanced Machine Learning and AI
scipy==1.11.1
transformers==4.35.2
datasets==2.14.5
spacy==3.7.2
faiss-cpu==1.7.4

# Graph Algorithms and Network Analysis
networkx==3.1
plotly==5.17.0

# Natural Language Processing
nltk>=3.8.0
gensim==4.3.2

# Data Processing and Analysis
dask==2023.8.0
pyarrow==12.0.1
openpyxl==3.1.2

# Advanced Database Features
elasticsearch==8.11.0
kafka-python==2.0.2

# Performance and Monitoring
prometheus-client>=0.16.0
psutil>=5.9.0

# Enterprise Security and Configuration
jwt>=1.3.1
python-decouple>=3.8

# Visualization and Reporting
seaborn==0.12.2
graphviz==0.20.1

# Configuration and Environment
pyyaml==6.0.1

# Enterprise Integrations
azure-storage-blob>=12.16.0
google-cloud-storage>=2.9.0
apache-airflow-client>=2.6.0

# CRITICAL: PyTorch for enterprise AI capabilities
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.2+cpu
sentence-transformers==2.2.2

# Enterprise AI and ML
firebase-admin>=6.2.0
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0